pyproject.toml
[project]
name = "portrecon"
version = "0.1.0"
description = "Port reconnaissance scanner with FastAPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-nmap>=0.7.1",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "rich>=13.7.0",
    "loguru>=0.7.2",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",
]
